+++ src/is/focusable.js
@@ -7,28 +7,9 @@
 //   Gecko: `svg a[xlink|href]` is not identified as focusable (because SVGElement.prototype.focus is missing)
 //   Blink, WebKit: SVGElements that have been made focusable by adding a focus event listener are not identified as focusable
 
-import '../prototype/svgelement.prototype.focus';
-import '../prototype/element.prototype.matches';
-import getParents from '../get/parents';
+import isFocusRelevant from './focus-relevant';
 import isVisible from './visible';
 import isDisabled from './disabled';
-import isValidTabindex from './valid-tabindex';
-import isValidArea from './valid-area';
-
-let canFocusSvgMethod = SVGElement.prototype.focus === HTMLElement.prototype.focus;
-import canFocusAudioWithoutControls from '../supports/focus-audio-without-controls';
-import canFocusFieldset from '../supports/focus-fieldset';
-import canFocusHtml from '../supports/focus-html';
-import canFocusImgIsmap from '../supports/focus-img-ismap';
-import canFocusImgUsemapTabindex from '../supports/focus-img-usemap-tabindex';
-import canFocusLabelTabindex from '../supports/focus-label-tabindex';
-import canFocusScrollBody from '../supports/focus-scroll-body';
-import canFocusScrollContainer from '../supports/focus-scroll-container';
-import canFocusScrollContainerWithoutOverflow from '../supports/focus-scroll-container-without-overflow';
-import canFocusSummary from '../supports/focus-summary';
-import canFocusSvg from '../supports/focus-svg';
-import canFocusTable from '../supports/focus-table';
-import canFocusVideoWithoutControls from '../supports/focus-video-without-controls';
 
 export default function(element) {
   if (element === document) {
@@ -39,22 +20,7 @@ export default function(element) {
     throw new TypeError('is/focusable requires an argument of type Element');
   }
 
-  const nodeName = element.nodeName.toLowerCase();
-
-  // object[usemap] is not focusable in any browser
-  if (nodeName === 'object' && element.hasAttribute('usemap')) {
-    return false;
-  }
-
-  // input[type="hidden"] cannot be focused
-  if (nodeName === 'input' && element.type === 'hidden') {
-    return false;
-  }
-
-  if (nodeName === 'label' && !canFocusLabelTabindex) {
-    // there's no way to make an element focusable other than by adding a tabindex,
-    // and focus behavior of the label element seems hard-wired to ignore tabindex
-    // in some browsers (like Gecko, Blink and WebKit)
+  if (!isFocusRelevant(element)) {
     return false;
   }
 
@@ -62,135 +28,10 @@ export default function(element) {
     return false;
   }
 
-  if (nodeName === 'area') {
-    return isValidArea(element);
-  }
-
   // elements that are not rendered, cannot be focused
   if (!isVisible(element)) {
     return false;
   }
 
-  if (nodeName === 'input' || nodeName === 'select' || nodeName === 'button' || nodeName === 'textarea') {
-    // disabled elements have been filtered out above, using isDisabled()
-    return true;
-  }
-
-  if (nodeName === 'a' && element.hasAttribute('href')) {
-    return true;
-  }
-
-  if (nodeName === 'iframe' || nodeName === 'object' || nodeName === 'embed') {
-    // browsing context containers
-    return true;
-  }
-
-  if (element.hasAttribute('contenteditable')) {
-    // also see CSS property user-modify below
-    return true;
-  }
-
-  if (nodeName === 'audio' && (canFocusAudioWithoutControls || element.hasAttribute('controls'))) {
-    return true;
-  }
-
-  if (nodeName === 'video' && (canFocusVideoWithoutControls || element.hasAttribute('controls'))) {
-    return true;
-  }
-
-  if (canFocusSummary && nodeName === 'summary') {
-    return true;
-  }
-
-  if (nodeName === 'img' && element.hasAttribute('usemap') && isValidTabindex(element)) {
-    // Gecko, Trident and Edge do not allow an image with an image map and tabindex to be focused,
-    // it appears the tabindex is overruled so focus is still forwarded to the <map>
-    return canFocusImgUsemapTabindex;
-  }
-
-  // TODO: not not identify body as focusable
-  if (nodeName === 'body') {
-    return true;
-  }
-
-  // TODO: not not identify html as focusable
-  if (canFocusHtml && nodeName === 'html') {
-    // Firefox, IE11 can focus <html>
-    return true;
-  }
-
-  if (canFocusTable && (nodeName === 'table' || nodeName === 'td')) {
-    // IE10-11 can focus <table> and <td>
-    return true;
-  }
-
-  if (canFocusFieldset && nodeName === 'fieldset') {
-    // IE10-11 can focus <fieldset>
-    return true;
-  }
-
-  if (canFocusSvgMethod && canFocusSvg && nodeName === 'svg') {
-    // NOTE: in Chrome this would be something like 'svg, svg *,' as *every* svg element with a focus event listener is focusable
-    return true;
-  }
-
-  if (canFocusSvgMethod && element.matches('svg a[*|href]')) {
-    // Namespace problems of [xlink:href] explained in http://stackoverflow.com/a/23047888/515124
-    // Firefox cannot focus <svg> child elements from script
-    return true;
-  }
-
-  // http://www.w3.org/TR/html5/editing.html#sequential-focus-navigation-and-the-tabindex-attribute
-  if (isValidTabindex(element)) {
-    return true;
-  }
-
-  const style = window.getComputedStyle(element, null);
-  const userModify = style.webkitUserModify || '';
-  if (userModify && userModify.indexOf('write') !== -1) {
-    // http://www.w3.org/TR/1999/WD-css3-userint-19990916#user-modify
-    // https://github.com/medialize/ally.js/issues/17
-    return true;
-  }
-
-  if (canFocusImgIsmap && nodeName === 'img' && element.hasAttribute('ismap')) {
-    // IE10-11 considers the <img> in <a href><img ismap> focusable
-    // https://github.com/medialize/ally.js/issues/20
-    let hasLinkParent = getParents({context: element}).some(
-      parent => parent.nodeName.toLowerCase() === 'a' && parent.hasAttribute('href')
-    );
-
-    if (hasLinkParent) {
-      return true;
-    }
-  }
-
-  if (canFocusScrollContainer && (element.offsetHeight < element.scrollHeight || element.offsetWidth < element.scrollWidth)) {
-    // scrollable containers are focusable Internet Explorer
-    // scrollable containers are tabbable in Firefox
-    // https://github.com/medialize/ally.js/issues/21
-    if (canFocusScrollContainerWithoutOverflow) {
-      return true;
-    }
-    // Firefox requires proper overflow setting, IE does not
-    // https://developer.mozilla.org/en-US/docs/Web/CSS/overflow
-    return [
-      style.getPropertyValue('overflow'),
-      style.getPropertyValue('overflow-x'),
-      style.getPropertyValue('overflow-y'),
-    ].some(overflow => overflow === 'auto' || overflow === 'scroll');
-  }
-
-  const parent = element.parentElement || element.parentNode;
-  if (canFocusScrollBody && (parent.offsetHeight < parent.scrollHeight || parent.offsetWidth < parent.scrollWidth)) {
-    // scrollable bodies are focusable Internet Explorer
-    // https://github.com/medialize/ally.js/issues/21
-    return true;
-  }
-
-  // NOTE: elements marked as inert are not focusable,
-  // but that property is not exposed to the DOM
-  // http://www.w3.org/TR/html5/editing.html#inert
-
-  return false;
+  return true;
 }